consumer/requirements.txt
paho-mqtt==2.1.0
pymongo==4.13.2